Move CSS to emphasize module entry points
index.js is the expected entry point for a module. Naming the CSS
to match was intended to emphasize its relationship but could be
a red herring for junior developers.

I'm also adding a TODO note that the CSS should probably be in a
template applied to the shadow root rather than global styles.